requirements.txt
pytest~=8.3.3
allure-pytest
selene>=2.0.0rc2
Appium-Python-Client~=4.3.0
python-dotenv==1.0.1
pydantic==2.5.2
requests~=2.31.0